View edge insets
[Add]
Add UIView+Insets for safe area insets, UIButton insets

[Remove]
Remove global extension

[Move]
Move safeAreaInsets to UIView+Insets
