+++ Marumaru/PlayMangaViewController.swift
@@ -44,7 +44,6 @@ class PlayMangaViewController: UIViewController {
     // MARK: - LifeCycle
     override func viewDidLoad() {
         super.viewDidLoad()
-        
         setup()
         bind()
         playManga()
