+++ Marumaru/SearchViewController.swift
@@ -29,7 +29,6 @@ class SearchViewController: UIViewController {
     // MARK: - LifeCycle
     override func viewDidLoad() {
         super.viewDidLoad()
-        
         setup()
         bind()
     }
